+++ client/src/api/downloadButton.jsx
@@ -1,28 +1,41 @@
 import { Button } from "@mui/material";
 
-export const DownloadFile = ({ filename, content, label }) => {
-  const downloadFile = () => {
-      // Create a blob with the content
-      const blob = new Blob([content], { type: "text/plain;charset=utf-8" });
+export const DownloadFile = ({ filename, content, contentGetter, label }) => {
+    const downloadFile = async () => {
+        // Get the content
+        if (contentGetter) {
+            try {
+                content = await contentGetter();
+                if(typeof content !== "string") {
+                    content = JSON.stringify(content, null, 2);
+                }
+            } catch (e) {
+                console.error(e);
+                return;
+            }
+        }
 
-      // Create a link element
-      const link = document.createElement("a");
-      link.href = URL.createObjectURL(blob);
-      link.download = filename;
+        // Create a blob with the content
+        const blob = new Blob([content], { type: "text/plain;charset=utf-8" });
 
-      // Append the link to the document (needed for Firefox)
-      document.body.appendChild(link);
+        // Create a link element
+        const link = document.createElement("a");
+        link.href = URL.createObjectURL(blob);
+        link.download = filename;
 
-      // Simulate a click to start the download
-      link.click();
+        // Append the link to the document (needed for Firefox)
+        document.body.appendChild(link);
 
-      // Cleanup the DOM by removing the link element
-      document.body.removeChild(link);
-  }
+        // Simulate a click to start the download
+        link.click();
 
-  return (
-      <Button onClick={downloadFile}>
-          {label}
-      </Button>
-  );
+        // Cleanup the DOM by removing the link element
+        document.body.removeChild(link);
+    }
+
+    return (
+        <Button onClick={downloadFile}>
+            {label}
+        </Button>
+    );
 }